requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
sqlalchemy==2.0.48
asyncpg==0.30.0
python-dotenv==1.2.2
